+++ db/migrate/20230217160251_add_involuntary_check_provisional_conversion_date.rb
@@ -1,0 +1,5 @@
+class AddInvoluntaryCheckProvisionalConversionDate < ActiveRecord::Migration[7.0]
+  def change
+    add_column :conversion_involuntary_task_lists, :stakeholder_kick_off_check_provisional_conversion_date, :boolean
+  end
+end